Fix coupling between scheduler and rescheduler by moving rescheduler code under notifier package where it's used and belongs.
