+++ st2actions/st2actions/scheduler.py
@@ -13,11 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import datetime
-
-from apscheduler.schedulers.background import BlockingScheduler
-from apscheduler.triggers.interval import IntervalTrigger
-import apscheduler.util as aps_utils
 from kombu import Connection
 from oslo.config import cfg
 
@@ -26,17 +21,14 @@ from st2common.constants import action as action_constants
 from st2common.exceptions.db import StackStormDBObjectNotFoundError
 from st2common.models.db.liveaction import LiveActionDB
 from st2common.services import action as action_service
-from st2common.services import coordination
 from st2common.persistence.liveaction import LiveAction
 from st2common.persistence.policy import Policy
 from st2common import policies
 from st2common.transport import consumers, liveaction
 from st2common.util import action_db as action_utils
-from st2common.util import date as date_utils
 
 __all__ = [
     'get_scheduler',
-    'get_rescheduler'
 ]
 
 
@@ -106,52 +98,3 @@ class ActionExecutionScheduler(consumers.MessageHandler):
 def get_scheduler():
     with Connection(cfg.CONF.messaging.url) as conn:
         return ActionExecutionScheduler(conn, [ACTIONRUNNER_REQUEST_Q])
-
-
-def recover_delayed_executions():
-    coordinator = coordination.get_coordinator()
-    dt_now = date_utils.get_datetime_utc_now()
-    dt_delta = datetime.timedelta(seconds=cfg.CONF.scheduler.delayed_execution_recovery)
-    dt_timeout = dt_now - dt_delta
-
-    with coordinator.get_lock('st2-rescheduling-delayed-executions'):
-        liveactions = LiveAction.query(status=action_constants.LIVEACTION_STATUS_DELAYED,
-                                       start_timestamp__lte=dt_timeout,
-                                       order_by=['start_timestamp'])
-
-        if not liveactions:
-            return
-
-        LOG.info('There are %d liveactions that have been delayed for longer than %d seconds.',
-                 len(liveactions), cfg.CONF.scheduler.delayed_execution_recovery)
-
-        # Update status to requested and publish status for each liveactions.
-        rescheduled = 0
-        for instance in liveactions:
-            try:
-                action_service.update_status(instance,
-                                             action_constants.LIVEACTION_STATUS_REQUESTED,
-                                             publish=True)
-                rescheduled += 1
-            except:
-                LOG.exception('Unable to reschedule liveaction. <LiveAction.id=%s>', instance.id)
-
-        LOG.info('Rescheduled %d out of %d delayed liveactions.', len(liveactions), rescheduled)
-
-
-def get_rescheduler():
-    timer = BlockingScheduler()
-
-    time_spec = {
-        'seconds': cfg.CONF.scheduler.rescheduling_interval,
-        'timezone': aps_utils.astimezone('UTC')
-    }
-
-    timer.add_job(recover_delayed_executions,
-                  trigger=IntervalTrigger(**time_spec),
-                  max_instances=1,
-                  misfire_grace_time=60,
-                  next_run_time=date_utils.get_datetime_utc_now(),
-                  replace_existing=True)
-
-    return timer
